requirements.txt
absl-py==1.4.0
accelerate==0.20.3
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.39.1
blessed==1.20.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
comm==0.1.3
contourpy==1.1.0
croniter==1.4.1
cycler==0.11.0
datasets==2.13.1
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.1
dill==0.3.6
docker-pycreds==0.4.0
evaluate==0.4.0
exceptiongroup==1.1.2
executing==1.2.0
fairscale==0.4.13
fastapi==0.100.0
fastjsonschema==2.17.1
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.15.1
idna==3.4
inquirer==3.1.3
ipykernel==6.24.0
ipython==8.14.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lightning==2.0.5
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==16.0.6
loralib==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
minepy==1.2.6
mlxtend==0.22.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ordered-set==4.1.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
peft @ git+https://github.com/huggingface/peft@032fff92fb74b737a2934e91a08d82142fb79dc3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
plotly==5.15.0
ppscore==1.3.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.5
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
readchar==4.0.5
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
responses==0.18.0
rich==13.4.2
rouge-score==0.1.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.26.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
-e git+https://github.com/Myashka/sollama.git@bdf41c0536400d59c35f2bb3cc618b6647dff321#egg=src
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers@4b26a61631b8fd30f845cf08ebcc5ed65fe83c9b
triton==2.0.0
trl==0.6.0
typing_extensions==4.7.1
tzdata==2023.3
UpSetPlot==0.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.0
wandb==0.15.5
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
xxhash==3.2.0
yarl==1.9.2
